backend-requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.51.0
ffmpeg-python==0.2.0
pyannote.audio==3.1.1
torch==2.0.1
librosa==0.10.1
scikit-learn==1.3.0
numpy<2.0
Werkzeug==2.3.7